requirements.txt
langchain_openai 
langchain_core
python-dotenv
streamlit
langchain_community
langserve
fastapi
uvicorn
sse_starlette
pypdf
chromadb
faiss-cpu
PyPDF2
fitz
elasticsearch
groq
langchain_groq

